add missing semicolons in tests
The tests were loading File::Spec, but were missing the trailing
semicolons. This resulted in the return value from the open call being
passed in to File::Spec->import. Perl would then ignore that call to a
nonexistent import. Future versions of perl are intending to make calls
to undefined import methods with arguments an error.